Fix compilation warnings.
git-svn-id: https://scm.mvapich.cse.ohio-state.edu/svn/mpi/mvapich2/trunk@3052 09bc9535-d30e-0410-b1f7-d46b20a4725c

+++ src/util/mem/trmem.c
@@ -254,7 +254,7 @@ void *MPIU_trmalloc( size_t a, int lineno, const char fname[] )
 	/* Note that %08p (what we'd like to use) isn't accepted by
 	   all compilers */
 	MPIU_Error_printf( "[%d] Allocating %d bytes at %8p in %s:%d\n", 
-		     world_rank, a, new, fname, lineno );
+		     world_rank, (int)a, new, fname, lineno );
     }
     return (void *)new;
 }
